requirements.txt
python-telegram-bot
python-dotenv
Pillow
qris-payment==1.1.3
